v2.64.2

+++ packages/mercenary-core/tasks/build.js
@@ -1,14 +1,14 @@
 const path = require('path');
 const cp = require('child_process');
 const startProd = require('./startProd');
-const buildStatic = require('./static');
+const buildStatic = require('./buildStatic');
 
 const cwd = process.cwd();
 const configDir = path.join(__dirname, '../config');
 const webpack = require.resolve('.bin/webpack');
 
 // Production build
-const build = (config = { silent: false }) => {
+const build = (config = { silent: false, static: false }) => {
   let output = '';
 
   if (config.silent) {
@@ -23,9 +23,20 @@ const build = (config = { silent: false }) => {
       "${configDir}/webpack/production.js" ${output}
   `, { stdio: 'inherit' });
 
-  const prod = startProd({ async: true });
-  buildStatic();
-  prod.kill('SIGINT');
+  if (config.static) {
+    // Start the production server in "static" mode, which bypasses some of the
+    // standard production settings (force WWW, etc.)
+    const prod = startProd({
+      async: true,
+      mode: 'static',
+    });
+
+    // Run the static builder
+    buildStatic();
+
+    // Kill the server
+    prod.kill('SIGINT');
+  }
 };
 
 module.exports = build;
